added jwx.go
