+++ pkg/jwx/jwx.go
@@ -1,0 +1,18 @@
+package jwx
+
+import (
+	"github.com/lestrrat-go/jwx/jwk"
+)
+
+type JWX interface {
+	CreateJWK() (*jwk.Key, error)
+	CreateJWS() ([]byte, error)
+	VerifyJWS()
+	EncryptJWE()
+	DecryptJWE()
+}
+
+type jwxImpl struct {
+	SigPair JWKSignaturePair
+	keyAlg
+}
